+++ release_notes.md
@@ -4,6 +4,10 @@
 
 -   APP-5036 - [Services] Update message broker connection to not set tls_version
 -   APP-5037 - [API] Updated V3 API to include new endpoints and/or updates to existing endpoints
+-   APP-5055 - [Services] Updated Paho MQTT library and message broker reconnect logic
+-   APP-5067 - [Services] Fix issue with redis connection on each new thread
+-   APP-5068 - [Batch] Added auto truncate of attribute values
+-   APP-5072 - [Services] Applied PR #356
 
 ## 4.0.10